Changes:
1. docs/Tests.Detection-of-unstable-tests.md: Description of unstable
tests detection procedure with results for mainnet branch on 2019-05-15 (commit a70894c8c4223424151cdff7441b1fb2e6bad309)
2. ci_scripts/run-internal-tests.sh, ci_scripts/run-pkg-tests.sh -
scripts created during detection
3. pkg/messaging/client_test.go, pkg/messaging/client_no_ci_test.go -
unstable test TestClientConnectInitialServers tagged with '!no_ci'
4. formatted, linted, checked
